Move journals into their parent source.
