+++ frontend/data.cpp
@@ -138,11 +138,15 @@ bool MainWindow::highlightRunNumber(int runNumber)
 
 void MainWindow::on_actionRefresh_triggered()
 {
-    if (!currentJournal_)
+    if (!currentJournalSource_)
         return;
-    auto &journal = currentJournal_->get();
+    auto &journalSource = currentJournalSource();
+    auto optJournal = journalSource.currentJournal();
+    if (!optJournal)
+        return;
 
-    backend_.getJournalUpdates(journal.location(), [=](HttpRequestWorker *worker) { handleGetJournalUpdates(worker); });
+    backend_.getJournalUpdates(optJournal->get().location(),
+                               [=](HttpRequestWorker *worker) { handleGetJournalUpdates(worker); });
 }
 
 // Jump to run number
