+++ frontend/journalSource.h
@@ -3,7 +3,12 @@
 
 #pragma once
 
+#include "journal.h"
+#include "optionalRef.h"
 #include <QString>
+
+// Forward Declarations
+class HttpRequestWorker;
 
 // Journal Source Definition
 class JournalSource
@@ -18,7 +23,7 @@ class JournalSource
     // Return text string for specified JournalSource type
     static QString journalSourceType(JournalSourceType type);
     // Convert text string to JournalSource type
-    static JournalSourceType journalSourceType(QString typeString);
+    static JournalSourceType journalSourceType(const QString &typeString);
     // Data Organisation Types
     enum class DataOrganisationType
     {
@@ -95,4 +100,27 @@ class JournalSource
     void setState(JournalSourceState state);
     // Return current state of the journal source
     JournalSourceState state() const;
+
+    /*
+     * Journals
+     */
+    private:
+    // Available journals
+    std::vector<Journal> journals_;
+    // Currently selected journal (if any)
+    OptionalReferenceWrapper<Journal> currentJournal_;
+
+    public:
+    // Clear current journals
+    void clearJournals();
+    // Add new journal
+    Journal &addJournal(const QString &name, const Locator &location);
+    // Return available journals
+    const std::vector<Journal> &journals() const;
+    // Find named journal
+    OptionalReferenceWrapper<Journal> findJournal(const QString &name);
+    // Set current journal being displayed
+    void setCurrentJournal(QString name);
+    // Return current journal
+    OptionalReferenceWrapper<Journal> currentJournal() const;
 };
